Extract Containers to separate component areas
